refactor: comply with GDScript style guide
I have rewritten my code to mostly comply with the Godot Doc's style
guide, mainly prefixing private variables with an underscore and code
order. One big exception is that I use 4 spaces instead of tabs for
better viewing in Github. In addition, I have elected to not prefix
public interface methods that are meant to be overriden with an
underscore. Also added Python-style docstrings and reduced the number of
magic constants. Per-level instructions have been added as well.