feat: add background lines
These serve as part eye candy, part aid for quickly comparing elements
that are close in size but far apart in merge sort. To achieve this,
ArrayView had to be refactored to no longer use a Viewport, so this also
removes the possibility of animating merge sort the way it was.